+++ course/format/flexsections/lib.php
@@ -25,6 +25,7 @@
 defined('MOODLE_INTERNAL') || die();
 require_once($CFG->dirroot. '/course/format/lib.php');
 
+use format_flexsections\constants;
 use core\output\inplace_editable;
 
 define('FORMAT_FLEXSECTIONS_COLLAPSED', 1);
@@ -54,7 +55,17 @@ class format_flexsections extends core_courseformat\base {
      * @return bool
      */
     public function uses_course_index() {
-        return true;
+        return $this->get_course_index_display() != constants::COURSEINDEX_NONE;
+    }
+
+    /**
+     * Type of course index display
+     *
+     * @return int
+     */
+    public function get_course_index_display(): int {
+        $type = (int)get_config('format_flexsections', 'courseindexdisplay');
+        return $type;
     }
 
     /**
@@ -74,6 +85,15 @@ class format_flexsections extends core_courseformat\base {
     public function get_max_section_depth(): int {
         $limit = (int)get_config('format_flexsections', 'maxsectiondepth');
         return max(1, min($limit, 100));
+    }
+
+    /**
+     * Accordion effect
+     *
+     * @return bool
+     */
+    public function get_accordion_setting(): bool {
+        return (bool)get_config('format_flexsections', 'accordion');
     }
 
     /**
@@ -192,11 +212,6 @@ class format_flexsections extends core_courseformat\base {
                 $url->param('section', $parent);
             }
             $url->set_anchor('section-'.$sectionno);
-            return $url;
-        } else {
-            // General section.
-            $url->set_anchor('section-'.$sectionno);
-            return $url;
         }
         return $url;
     }
@@ -413,6 +428,39 @@ class format_flexsections extends core_courseformat\base {
                 'default' => COURSE_DISPLAY_SINGLEPAGE,
             )
         );
+    }
+
+    /**
+     * Definitions of the additional options that this course format uses for course.
+     *
+     * Flexsections format uses the following options:
+     * - showsection0title
+     *
+     * @param bool $foreditform
+     * @return array of options
+     */
+    public function course_format_options($foreditform = false) {
+        static $courseformatoptions = false;
+        if ($courseformatoptions === false) {
+            $courseformatoptions = [
+                'showsection0title' => [
+                    'default' => get_config('format_flexsections', 'showsection0titledefault') ?? 0,
+                    'type' => PARAM_BOOL,
+                ],
+            ];
+        }
+        if ($foreditform && !isset($courseformatoptions['showsection0title']['label'])) {
+            $courseformatoptionsedit = [
+                'showsection0title' => [
+                    'label' => new lang_string('showsection0title', 'format_flexsections'),
+                    'help' => 'showsection0title',
+                    'help_component' => 'format_flexsections',
+                    'element_type' => 'advcheckbox',
+                ],
+            ];
+            $courseformatoptions = array_merge_recursive($courseformatoptions, $courseformatoptionsedit);
+        }
+        return $courseformatoptions;
     }
 
     /**
@@ -1256,6 +1304,44 @@ class format_flexsections extends core_courseformat\base {
         // Partial rebuild section cache that has been purged.
         rebuild_course_cache($this->courseid, true, true);
     }
+
+    /**
+     * Display 'Add section' as a link on the page and not as a "Add subsection" menu item
+     *
+     * @param int $sectionnum
+     * @return bool
+     */
+    public function should_display_add_sub_section_link(int $sectionnum): bool {
+        // Display for the top-level sections and for the sections that are displayed as a link.
+        if (!$sectionnum) {
+            return true;
+        }
+        $section = $this->get_section($sectionnum);
+        return (bool)$section->collapsed;
+    }
+
+    /**
+     * Method used to get the maximum number of sections for this course format.
+     *
+     * Flexsections does not have a limit for the total number of the sections.
+     *
+     * @return int
+     */
+    public function get_max_sections() {
+        return 9999999;
+    }
+
+    /**
+     * Method used to get the maximum number of sections on the top level.
+     * @return int
+     */
+    public function get_max_toplevel_sections() {
+        $maxsections = get_config('moodlecourse', 'maxsections');
+        if (!isset($maxsections) || !is_numeric($maxsections)) {
+            $maxsections = 52;
+        }
+        return $maxsections;
+    }
 }
 
 /**
@@ -1285,4 +1371,43 @@ function format_flexsections_get_fontawesome_icon_map() {
     return [
         'format_flexsections:mergeup' => 'fa-level-up',
     ];
+}
+
+/**
+ * If we are on an activity page inside the course in the 'flexsections' format - return the activity
+ *
+ * @return cm_info|null
+ */
+function format_flexsections_add_back_link_to_cm(): ?cm_info {
+    global $PAGE, $CFG;
+    if ($PAGE->course
+            && $PAGE->cm
+            && $PAGE->course->format === 'flexsections' // Only modules in 'flexsections' courses.
+            && get_config('format_flexsections', 'cmbacklink')
+            && $PAGE->pagelayout === 'incourse' // Only view pages with the incourse layout (not popup, embedded, etc).
+            && $PAGE->cm->sectionnum // Do not display in activities in General section.
+            && $PAGE->url->out_omit_querystring() === $CFG->wwwroot . "/mod/{$PAGE->cm->modname}/view.php") {
+        return $PAGE->cm;
+    }
+    return null;
+}
+
+/**
+ * Callback allowing to add contetnt inside the region-main, in the very end
+ *
+ * If we are on activity page, add the "Back to section" link
+ *
+ * @return string
+ */
+function format_flexsections_before_footer() {
+    global $OUTPUT;
+    if ($cm = format_flexsections_add_back_link_to_cm()) {
+        return $OUTPUT->render_from_template('format_flexsections/back_link_in_cms', [
+            'backtosection' => [
+                'url' => course_get_url($cm->course, $cm->sectionnum)->out(false),
+                'sectionname' => get_section_name($cm->course, $cm->sectionnum),
+            ]
+        ]);
+    }
+    return '';
 }