+++ course/format/flexsections/CHANGELOG.md
@@ -1,6 +1,25 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [4.0.4] - 2023-07-02
+### Added
+- Setting to show header for the General section and make it collapsible
+- Link 'Add section' between sections (only for the top-level sections on the
+  current page)
+- Setting how to show the course index (sections and activities, only sections,
+  do not display)
+- Setting to enable 'accordion' effect - when one section is expanded, all others
+  are collapsed
+- Setting to display 'Back to...' link inside the activities allowing to return
+  back to the course section
+- Setting 'maxsections' now only affects the number of top-level sections. Number
+  of subsections is unlimited (there is also a setting about the maximum depth).
+### Fixed
+- Fixed a bug when drag&drop of activities was not possible if the target
+  section is empty
+
+Release contributor: **Hogeschool Inholland, the Netherlands**.
+
 ## [4.0.3] - 2023-05-06
 ### Added
 - Allow to indent activities on the course page